renamed new modules, removing new_ prefix
